Revert "[김유정]"
This reverts commit 68cf475dd2ac33715972b8727f42c5c385cd27dc.

+++ src/test/java/com/Bhealthy/sympathy/bo/SympathyTest.java
@@ -25,10 +25,10 @@ class SympathyTest {
 	@Autowired
 	private SympathyMapper sympathyMapper;
 	
-	//@Test
-	@DisplayName("공감하기 추가 테스트")
-	void 공감하기추가() {
-		log.info("#####공감하기 추가 테스트");
+	@Test
+	@DisplayName("공감하기 추가/삭제 테스트")
+	void 공감하기() {
+		log.info("#####공감하기 추가/삭제 테스트");
 		//given
 		Sympathy sympathy = new Sympathy();
 		
@@ -41,28 +41,15 @@ class SympathyTest {
 		// then
 		log.info("공감하기" + sympathyCount);
 		// 공감 존재 -> 삭제
-		assertTrue(sympathyCount > 0);
-		sympathyMapper.deleteSympathyByPostIdOrUserId(sympathy.getPostId(), sympathy.getUserId());
-		log.info("#####공감하기 삭제 완료" + sympathyCount);
-				
+		if(sympathyCount > 0) {
+			sympathyMapper.deleteSympathyByPostIdOrUserId(sympathy.getPostId(), sympathy.getUserId());
+			log.info("#####공감하기 삭제 완료" + sympathyCount);
+		} else { // 공감 없음 -> 추가
+			sympathyMapper.insertSympathy(sympathy.getPostId(), sympathy.getUserId());
+			log.info("#####공감하기 추가 완료" + sympathyCount);
+		}
+	
+		
 	}
-	@Test
-	@DisplayName("공감하기 삭제 테스트")
-	void 공감하기삭제() {
-		log.info("#####공감하기 삭제 테스트");
-		//given X
-		Sympathy sympathy = new Sympathy();
-		
-		sympathy.setPostId(5);
-		sympathy.setUserId(5);
-		// when
-		int sympathyCount = sympathyMapper.selectSympathyCountByPostOrIdUserId(sympathy.getPostId(), sympathy.getUserId());
-		
-		// then
-		log.info("공감하기" + sympathyCount);
-		
-		assertNotEquals(sympathyCount, 1);// 공감 없음 -> 추가
-		sympathyMapper.insertSympathy(sympathy.getPostId(), sympathy.getUserId());
-		log.info("#####공감하기 추가 완료" + sympathyCount);
-	}
+
 }
